core: Inherit the bool module
